RIP old code, hello new codebase

+++ edge.hpp
@@ -5,9 +5,8 @@ struct Edge
 {
 	std::string start;
 	std::string end;
-	double weight;
-	//bool traversed = false;
-	Edge(const std::string& start, const double& weight, const std::string& end)
+	long double weight;
+	Edge(const std::string& start, const long double& weight, const std::string& end)
 		: start(start), end(end), weight(weight)
 	{};
 };
